text_analyser/requirements.txt
grpcio
protobuf
prometheus_client